% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlacklistGroups.R
\name{BlacklistGroups}
\alias{BlacklistGroups}
\title{Blacklist Groups}
\usage{
BlacklistGroups(data, groups, rhop, centroid_flag)
}
\arguments{
\item{data}{Processed data from CleanUpInput (or RemoveCellCycle)}

\item{groups}{Processed groups file from CleanUpInput}

\item{rhop}{x in mean+x*SD to determine upper cutoff for correlation in the blacklist. Default is 1.}

\item{centroid_flag}{use centroids if data is sparse}
}
\value{
newMedoids - new medoids data.frame for the new combined blacklisted clusters.

newGroups - new groups file containing cluster assignment based on new combined blacklisted clusters.
}
\description{
This function is used for the calculation of blacklisted medoids and groups. It calculates medoids, performs medoid correlations and creates the binary correlation table between medoids, otherwise known as the blacklist. It makes a blacklist heatmap. It makes new medoids based on blacklisted combined clusters and a new groups file based on the blacklisted combined clusters.
}
\keyword{blacklist}
\keyword{correlation}
