% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsDoublet.R
\name{IsDoublet}
\alias{IsDoublet}
\title{Is A Doublet}
\usage{
IsDoublet(data, newMedoids, groups, synthProfiles, log_file_name)
}
\arguments{
\item{data}{Processed data from CleanUpInput (or RemoveCellCycle).}

\item{newMedoids}{New combined medoids from BlacklistGroups.}

\item{groups}{Processed groups file from CleanUpInput.}

\item{synthProfiles}{Average profiles of synthetic doublets from SyntheticDoublets.}

\item{log_file_name}{used for saving run notes to log file}
}
\value{
isADoublet - data.frame with each cell as a row and whether it is called a doublet by deconvolution analysis.

resultsreadable - data.frame with results of deconvolution analysis (cell by cluster) in percentages.
}
\description{
This function uses deconvolution analysis (DeconRNASeq) to evaluate each cell for equal contribution from blacklisted clusters.
}
\keyword{decon}
\keyword{deconvolution}
\keyword{doublet}
