% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoMarkerFinder.R
\name{PseudoMarkerFinder}
\alias{PseudoMarkerFinder}
\title{Pseudo MarkerFinder}
\usage{
PseudoMarkerFinder(groups, redu_data2, full_data2, min_uniq = 4)
}
\arguments{
\item{groups}{Processed groups file from CleanUpInput.}

\item{redu_data2}{Processed data from CleanUpInput (or RemoveCellCycle) path, automatically written.}

\item{full_data2}{cleaned full expression matrix from CleanUpInput.}

\item{min_uniq}{minimum number of unique genes required for a cluster to be rescued}

\item{log_file_name}{used for saving run notes to log file}
}
\value{
new_table - non-doublet clusters, as determined by the "Remove" and "Rescue" steps.
}
\description{
This function uses ANOVA to look for unique gene expression in each possible doublet cluster.
}
\keyword{ANOVA}
\keyword{Finder}
\keyword{Marker}
