% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recluster.R
\name{Recluster}
\alias{Recluster}
\title{Recluster}
\usage{
Recluster(isADoublet, data, groups)
}
\arguments{
\item{isADoublet}{isADoublet data.frame from IsDoublet.}

\item{data}{Processed data from CleanUpInput (or RemoveCellCycle).}

\item{groups}{Processed groups file from CleanUpInput.}
}
\value{
newData2 - processed expression and groups file, reordered, with correct new cluster numbers.

decon - DeconCalledFreq table with all doublets 100 percent doublet and all non doublets at 0 percent doublet frequency.
}
\description{
This function reclusters doublets and non doublets seperately based on deconvolution analysis and returns a new expression file, groups file, and DeconFreq table for downstream analyses.
}
\keyword{recluster}
